create utils module
